refactor: rename surface/geom.rs to surface/geometry.rs
