+++ src/surface.rs
@@ -1,9 +1,9 @@
 //! Shared terminal-surface identity: the `OzmaTerminal` marker and the
 //! render-bundle injection observer, which fire for every surface — tmux
 //! panes and the Default-mode shell alike. Surface geometry helpers live in
-//! `geom`.
+//! `geometry`.
 
-pub(crate) mod geom;
+pub(crate) mod geometry;
 
 use bevy::prelude::*;
 use ozma_tty_renderer::material::TerminalUiMaterial;
